+++ main.ts
@@ -85,8 +85,6 @@ export default class PugTemplatePlugin extends Plugin {
     this.addSettingTab(new SettingTab(this.app, this));
   }
 
-  onunload() {}
-
   async loadSettings() {
     this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData());
   }