+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-08-20
+
 ### Added
 
 - Test page for attributes.
@@ -122,7 +124,9 @@ and this project adheres to
 - Default settings code
 
 [unreleased]:
-  https://github.com/nicholas-wilcox/pug-templates-obsidian-plugin/compare/0.3.3...HEAD
+  https://github.com/nicholas-wilcox/pug-templates-obsidian-plugin/compare/0.4.0...HEAD
+[0.4.0]:
+  https://github.com/nicholas-wilcox/pug-templates-obsidian-plugin/compare/0.3.3...0.4.0
 [0.3.3]:
   https://github.com/nicholas-wilcox/pug-templates-obsidian-plugin/compare/0.3.2...0.3.3
 [0.3.2]: